__app__/requirements.txt
azure-functions>=1.2.0
azure-functions-durable>=1.0.0b2